+++ src/app/clarity/actions.ts
@@ -1,0 +1,4 @@
+import { createAction } from 'typesafe-actions';
+import { ClarityDescription } from './descriptions/descriptionInterface';
+
+export const loadDescriptions = createAction('loadDescriptions')<ClarityDescription>();