+++ src/app/search/autocomplete.ts
@@ -3,6 +3,7 @@ import { t } from 'app/i18next-t';
 import { chainComparator, compareBy, reverseComparator } from 'app/utils/comparators';
 import { uniqBy } from 'app/utils/util';
 import _ from 'lodash';
+import { makeCommentString, parseQuery } from './query-parser';
 import { SearchConfig } from './search-config';
 import freeformFilters from './search-filters/freeform';
 
@@ -21,15 +22,28 @@ export const enum SearchItemType {
   // TODO: add types for exact-match item or perk that adds them to the query?
 }
 
+export interface SearchQuery {
+  /** The full text of the query */
+  fullText: string;
+  /** The query's top-level comment */
+  header?: string;
+  /** The query text excluding the top-level comment */
+  body: string;
+  /** Help text */
+  helpText?: string;
+}
+
 /** An item in the search autocompleter */
 export interface SearchItem {
   type: SearchItemType;
   /** The suggested query */
-  query: string;
+  query: SearchQuery;
   /** An optional part of the query that will be highlighted */
-  highlightRange?: [number, number];
-  /** Help text */
-  helpText?: React.ReactNode;
+  highlightRange?: {
+    section: 'header' | 'body';
+    /** The indices of the first and last character that should be highlighted */
+    range: [number, number];
+  };
 }
 
 /** matches a keyword that's probably a math comparison */
@@ -63,10 +77,13 @@ export default function createAutocompleter(searchConfig: SearchConfig) {
 
   return (query: string, caretIndex: number, recentSearches: Search[]): SearchItem[] => {
     // If there's a query, it's always the first entry
-    const queryItem = query
+    const queryItem: SearchItem | undefined = query
       ? {
           type: SearchItemType.Autocomplete,
-          query: query,
+          query: {
+            fullText: query,
+            body: query,
+          },
         }
       : undefined;
     // Generate completions of the current search
@@ -82,15 +99,22 @@ export default function createAutocompleter(searchConfig: SearchConfig) {
 
     // Help is always last...
     // Add an item for opening the filter help
-    const helpItem = {
+    const helpItem: SearchItem = {
       type: SearchItemType.Help,
-      query: query || '', // use query as the text so we don't change text when selecting it
+      query: {
+        // use query as the text so we don't change text when selecting it
+        fullText: query || '',
+        body: query || '',
+      },
     };
 
     // mix them together
     return [
       ..._.take(
-        uniqBy(_.compact([queryItem, ...filterSuggestions, ...recentSearchItems]), (i) => i.query),
+        uniqBy(
+          _.compact([queryItem, ...filterSuggestions, ...recentSearchItems]),
+          (i) => i.query.fullText
+        ),
         7
       ),
       helpItem,
@@ -152,14 +176,46 @@ export function filterSortRecentSearches(query: string, recentSearches: Search[]
   const recentSearchesForQuery = query
     ? recentSearches.filter((s) => s.query.includes(query))
     : Array.from(recentSearches);
-  return recentSearchesForQuery.sort(recentSearchComparator).map((s) => ({
-    type: s.saved
-      ? SearchItemType.Saved
-      : s.usageCount > 0
-      ? SearchItemType.Recent
-      : SearchItemType.Suggested,
-    query: s.query,
-  }));
+  return recentSearchesForQuery.sort(recentSearchComparator).map((s) => {
+    const ast = parseQuery(s.query);
+    const topLevelComment = ast.comment && makeCommentString(ast.comment);
+    const result: SearchItem = {
+      type: s.saved
+        ? SearchItemType.Saved
+        : s.usageCount > 0
+        ? SearchItemType.Recent
+        : SearchItemType.Suggested,
+      query: {
+        fullText: s.query,
+        header: ast.comment,
+        body: topLevelComment ? s.query.substring(topLevelComment.length).trim() : s.query,
+      },
+    };
+
+    // highlight the matched range of the query
+    if (query) {
+      if (result.query.header) {
+        const index = result.query.header.indexOf(query);
+        if (index !== -1) {
+          result.highlightRange = {
+            section: 'header',
+            range: [index, index + query.length],
+          };
+        }
+      }
+      if (!result.highlightRange) {
+        const index = result.query.body.indexOf(query);
+        if (index !== -1) {
+          result.highlightRange = {
+            section: 'body',
+            range: [index, index + query.length],
+          };
+        }
+      }
+    }
+
+    return result;
+  });
 }
 
 const caretEndRegex = /([\s)]|$)/;
@@ -203,15 +259,21 @@ export function autocompleteTermSuggestions(
     const filterDef = findFilter(word, searchConfig);
     const newQuery = base + word + query.slice(caretIndex);
     return {
-      query: newQuery,
+      query: {
+        fullText: newQuery,
+        body: newQuery,
+        helpText: filterDef
+          ? (Array.isArray(filterDef.description)
+              ? t(...filterDef.description)
+              : t(filterDef.description)
+            )?.replace(/\.$/, '')
+          : undefined,
+      },
       type: SearchItemType.Autocomplete,
-      highlightRange: [match.index, match.index + word.length],
-      helpText: filterDef
-        ? (Array.isArray(filterDef.description)
-            ? t(...filterDef.description)
-            : t(filterDef.description)
-          )?.replace(/\.$/, '')
-        : undefined,
+      highlightRange: {
+        section: 'body',
+        range: [match.index, match.index + word.length],
+      },
     };
   });
 }