+++ docs/CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Next
 
+* Weapon perks now include community-sourced weapon and armor perk descriptions courtesy of [Clarity](https://d2clarity.page.link/websiteDIM) and [Pip1n's Destiny Data Compendium](https://docs.google.com/spreadsheets/d/1WaxvbLx7UoSZaBqdFr1u32F2uWVLo-CJunJB4nlGUE4/htmlview?pru=AAABe9E7ngw*TxEsfbPsk5ukmr0FbZfK8w#). These can be disabled in settings.
+* When making automatic moves, DIM will always avoid filling in your last open Consumables slot. An item can still be manually moved into your character's pockets as the 50th consumable.
 * Loadout Optimizer will now suggest class items with an elemental affinity matching the mods even when allowing changes to elemental affinity.
+
+## 7.20.1 <span class="changelog-date">(2022-06-06)</span>
+
+* Fixed some items showing the wrong popup.
+
+## 7.20.0 <span class="changelog-date">(2022-06-05)</span>
+
+* The top level comment of a saved search filter is now displayed separately from the filter query.
+* Support for new loot: `source:duality` and `source:haunted`.
+* Little clearer warning when you have hidden a major section of your inventory.
+* Moved the currencies (glimmer, legendary shards, etc) from the "Armor" tab to the "Inventory" tab on mobile, and also included them in the material counts sheet (accessible from Vault header dropdown).
 
 ## 7.19.0 <span class="changelog-date">(2022-05-29)</span>